+++ Directory.Build.props
@@ -4,7 +4,7 @@
         <TargetFrameworks>net6.0;net8.0;</TargetFrameworks>
 
         <LangVersion>latest</LangVersion>
-        <Authors>Nimvox</Authors>
+        <Authors>ZyphCare</Authors>
         <Copyright>2024</Copyright>
         <RepositoryType>git</RepositoryType>
