--- src/Nimvox.EntityFramework.Common/Entities/Entity.cs
@@ -1,12 +1,0 @@
-﻿namespace Nimvox.EntityFramework.Common.Entities;
-
-/// <summary>
-///     Represents an entity.
-/// </summary>
-public abstract class Entity
-{
-    /// <summary>
-    ///     Gets or sets the ID of this entity.
-    /// </summary>
-    public string Id { get; set; } = default!;
-}
