--- src/Nimvox.EntityFramework.Common/Entities/OrderDirections.cs
@@ -1,16 +1,0 @@
-﻿namespace Nimvox.EntityFramework.Common.Entities;
-
-/// <summary>
-///     The order direction.
-/// </summary>
-public enum OrderDirection
-{
-    /// <summary>
-    ///     Ascending order.
-    /// </summary>
-    Ascending,
-
-    /// <summary>
-    /// </summary>
-    Descending
-}
