--- src/Nimvox.EntityFramework.Common/Extensions/DependancyInjectionExtensions.cs
@@ -1,16 +1,0 @@
-﻿using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.DependencyInjection.Extensions;
-using Nimvox.EntityFramework.Common.Services;
-
-namespace Nimvox.EntityFramework.Common.Extensions;
-
-public static class DependancyInjectionExtensions
-{
-    public static IServiceCollection AddMemoryStore<TEntity, TStore>(this IServiceCollection services)
-        where TStore : class
-    {
-        services.TryAddSingleton<MemoryStore<TEntity>>();
-        services.TryAddScoped<TStore>();
-        return services;
-    }
-}
