+++ src/ZyphCare.EntityFramework.Common/Entities/Entity.cs
@@ -1,0 +1,12 @@
+﻿namespace ZyphCare.EntityFramework.Common.Entities;
+
+/// <summary>
+///     Represents an entity.
+/// </summary>
+public abstract class Entity
+{
+    /// <summary>
+    ///     Gets or sets the ID of this entity.
+    /// </summary>
+    public string Id { get; set; } = default!;
+}
