+++ src/ZyphCare.EntityFramework.Common/Entities/OrderDirections.cs
@@ -1,0 +1,16 @@
+﻿namespace ZyphCare.EntityFramework.Common.Entities;
+
+/// <summary>
+///     The order direction.
+/// </summary>
+public enum OrderDirection
+{
+    /// <summary>
+    ///     Ascending order.
+    /// </summary>
+    Ascending,
+
+    /// <summary>
+    /// </summary>
+    Descending
+}
